refactor: 도메인 패키지 재분류
