+++ src/main/java/domain/board/BoardFactory.java
@@ -1,8 +1,8 @@
 package domain.board;
 
-import domain.Position;
+import domain.position.Position;
 import domain.piece.Piece;
-import domain.piece.Team;
+import domain.game.Team;
 import java.util.Map;
 
 public interface BoardFactory {
