+++ src/main/java/domain/piece/ActivePiece.java
@@ -1,9 +1,7 @@
-package domain.activePiece;
+package domain.piece;
 
-import domain.Position;
-import domain.piece.Piece;
-import domain.piece.PieceType;
-import domain.piece.Team;
+import domain.position.Position;
+import domain.game.Team;
 import java.util.List;
 
 public abstract class ActivePiece implements Piece {
