+++ src/main/java/domain/piece/Cannon.java
@@ -1,15 +1,13 @@
-package domain.activePiece;
+package domain.piece;
 
-
-import domain.Position;
-import domain.piece.PieceType;
-import domain.piece.Team;
+import domain.position.Position;
+import domain.game.Team;
 import java.util.List;
 
-public class Chariot extends ActivePiece {
+public class Cannon extends ActivePiece {
 
-    public Chariot(Team team) {
-        super(team, PieceType.CHA);
+    public Cannon(Team team) {
+        super(team, PieceType.PHO);
     }
 
     @Override
