+++ src/main/java/domain/piece/Piece.java
@@ -1,6 +1,6 @@
 package domain.piece;
 
-import domain.Position;
+import domain.position.Position;
 
 public interface Piece {
     boolean canMove(Position source, Position target);
