+++ src/test/java/domain/PositionTest.java
@@ -3,6 +3,7 @@ package domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import domain.position.Position;
 import org.junit.jupiter.api.Test;
 
 class PositionTest {
